+++ service/src/main/kotlin/io/provenance/explorer/web/v2/BlockController.kt
@@ -1,8 +1,6 @@
 package io.provenance.explorer.web.v2
 
-import io.provenance.explorer.domain.models.explorer.BlockDetail
-import io.provenance.explorer.domain.models.explorer.PagedResults
-import io.provenance.explorer.domain.models.explorer.RecentBlock
+import io.provenance.explorer.domain.models.explorer.BlockSummary
 import io.provenance.explorer.service.ExplorerService
 import io.provenance.explorer.web.BaseController
 import io.swagger.annotations.Api
@@ -25,18 +23,15 @@ class BlockController(private val explorerService: ExplorerService) : BaseContro
 
     @ApiOperation("Return block at current height")
     @GetMapping("/height")
-    fun blockHeight(): ResponseEntity<BlockDetail> = ResponseEntity.ok(explorerService.getBlockAtHeight(null))
+    fun blockHeight() = ResponseEntity.ok(explorerService.getBlockAtHeight(null))
 
     @ApiOperation("Return block at specified height")
     @GetMapping("/height/{height}")
-    fun blockHeight(@PathVariable height: Int):
-        ResponseEntity<BlockDetail> = ResponseEntity.ok(explorerService.getBlockAtHeight(height))
+    fun blockHeight(@PathVariable height: Int) = ResponseEntity.ok(explorerService.getBlockAtHeight(height))
 
     @ApiOperation("Returns most recent blocks")
     @GetMapping("/recent")
     fun recentBlocks(@RequestParam(required = true, defaultValue = "10") @Min(1) count: Int,
-                     @RequestParam(required = true, defaultValue = "1") @Min(1) page: Int,
-                     @RequestParam(required = false, defaultValue = "desc") sort: String):
-        ResponseEntity<PagedResults<RecentBlock>> =
-        ResponseEntity.ok(explorerService.getRecentBlocks(count, page - 1, sort))
+                     @RequestParam(required = true, defaultValue = "1") @Min(1) page: Int
+    ) = ResponseEntity.ok(explorerService.getRecentBlocks(count, page - 1))
 }